+++ Application/Commands/UpdateSeekCommand.cs
@@ -6,7 +6,7 @@ using Reezer.Domain.Utils;
 
 namespace Reezer.Application.Commands;
 
-public record UpdateSeekCommand(string RoomCode, double Position)
+public record UpdateSeekCommand(string RoomCode, long Position)
     : IRequest<OneOf<Success, NotFound>>;
 
 public class UpdateSeekHandler(IMusicRoomRepository roomRepository, IPublisher publisher)
